fix: up key working correct
